Add test database to database.yml
Note that we are using the same database for the `production` and `test` environments.
We are changing this in a upcoming PR where we dynamically define the database name using an `environment` variable.

Until we merge this other PR it should not cause conflicts with the production setup as the test infraestrcture is not triggered. Nevertheless it is something that should be resolved promptly.