Add self signed certificate
This is used for Travis and for temporary installations without a domain name.

With this commit users should be able to access https://ip_address.

Note that an alternative implementation would be to use a Pebble[1] server, which is a LetsEncrypt server used for testing purposes. However documenation is sparse and it has been complicated to setup.

Using a self signed ssl certificate is not ideal but it's a start.

[1] https://github.com/letsencrypt/pebble